Compute Today sections once per layout pass
contentSections re-filtered the podcast/video lists four to six times
per body evaluation, and the attribution footer re-parsed its Markdown
each pass; the footer is now its own view with a static string.
